updates to the code
I think I got it.  I modified the incoming serial event and the processing serial.  I also modified the ESPNOW recieving to have 2 if statements.  This way I was able to specify if it was a broadcast or not, as well as removed the resetting of the boolean that tracks espnow broadcast messages.  this fixed the issue of the espnow messages getting set back on the chanied commands